refactor: Relocate `DayOfWeek.kt` to util package (#49)
